requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
bcrypt
python-dotenv
jinja2
python-multipart
passlib[bcrypt]
python-jose
email-validator
mysql-connector-python